Updated guards and generalized some types

+++ src/utils.c
@@ -7,7 +7,7 @@
 int
 euclidean_distance(struct client *a, struct client *b)
 {
-    int16_t x_diff, y_diff;
+    int x_diff, y_diff;
     x_diff = a->geom.x - b->geom.x;
     y_diff = a->geom.y - b->geom.y;
     return pow(x_diff, 2) + pow(y_diff, 2);
